refactor: improve docs generation
