+++ docs/swagger.ts
@@ -1,11 +1,11 @@
 import swaggerAutogen from 'swagger-autogen'
-import { env } from './env'
+import { env } from '../src/env'
 
 const doc = {
   info: {
     version: 'v1.0.0',
-    title: 'Swagger Demo Project',
-    description: 'Implementation of Swagger with TypeScript',
+    title: 'Orange Portfólio API',
+    description: 'A API mais vitaminada do mundo!',
   },
   servers: [
     {
@@ -23,7 +23,7 @@ const doc = {
   },
 }
 
-const outputFile = './swagger_output.json'
+const outputFile = '../docs/swagger_output.json'
 const endpointsFiles = ['./src/app.ts']
 
 swaggerAutogen(outputFile, endpointsFiles, doc)
